Clarify error message when Prometheus data dir finds unexpected files
While trying to work with remote storage, I kept hitting the error
referenced on line 161. I assumed that the remote volume was empty, but ended
up finding a test file when I mounted it manually. The current error
does mention that it is potentially necessary to wipe the storage, but
I thought presenting an additional note about the potential number of
files might clarify what an operator needs to remove.